Replace game url parsing logic with DailyBoxScoresPage class (#116)
Replace existing game url parsing logic with `DailyBoxScoresPage` class and a `game_url_paths` method.
